refactoring distill
This huge commit moves distill rules from two locations into the
rialto_airflow.distiller module:

- rialto_airflow.harvest.distill
- rialto_airflow.publish.publication

The reason for this was to consolidate the rules in one place, and to
partition the large modules into submodules that were named for what
property the rules were addressing.

The tests were also moved from test/publish/test_publication_by_authors.py
and test/harvest/test_distill.py into relevant Python files in
test/distiller.
